requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
huggingface_hub==0.20.3
python-telegram-bot==20.7
feedparser==6.0.10
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
apscheduler==3.10.4
